Apply Clippy fixes

+++ crates/hyperqueue/src/server/state.rs
@@ -161,7 +161,7 @@ impl State {
 
     pub fn new_task_id(&mut self, task_count: JobTaskCount) -> TakoTaskId {
         let id = self.task_id_counter;
-        self.task_id_counter += task_count as u32;
+        self.task_id_counter += task_count;
         id.into()
     }
